Expose our own Balance struct
